[refactor] 주석 수정 / filter chain 설정 파일 옮김 / .gitignore 수정
