+++ auth-module/.gitignore
@@ -1,1 +1,1 @@
-!src/main/resources/security.yml
+!src/main/resources/application.yml
