+++ auth-module/src/main/java/com/inhabas/api/auth/CorsConfig.java
@@ -1,13 +1,16 @@
-package com.inhabas.api.config;
+package com.inhabas.api.auth;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
+/**
+ * 개발 서버에 한해서만 cors 를 적용함. 대신 개발 전용 api 에 대해 높은 수준의 access control 이 보장되어야 함.
+ */
 @Configuration
 @Profile("dev")
-public class WebSecurityCors_dev implements WebMvcConfigurer {
+public class CorsConfig implements WebMvcConfigurer {
 
     @Override
     public void addCorsMappings(CorsRegistry registry) {
