+++ auth-module/src/main/resources/application.yml
